+++ packages/plugin-e2e/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v1.6.0 (Mon Jun 24 2024)
+
+#### 🚀 Enhancement
+
+- Plugin E2E: Interact with Grafana http api on behalf of logged in user [#965](https://github.com/grafana/plugin-tools/pull/965) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
 # v1.5.0 (Tue Jun 11 2024)
 
 #### 🚀 Enhancement